Use real API
Use Faraday not HTTP to download from s3.  We're already using Faraday
as a dependency of JsonApiClient.  There is no need to bring in another
net/http wrapper.

The logger for JsonApiClient needs to be added to the connection and not
on to Faraday to avoid a warning about middleware and adapters.